Initial implementation of intelligent-tm
Summary:
We'd like to provide more intelligent suggestions to translators who are
working on perseus exercises.  In this initial commit we provide methods
for:
- grouping similar msgId strings
- creating templates based on one msgId and one msgStr
- creating translations from a template and different msgIds within the same grouping

There is also a regression test which compares translations created using
the template against existing translations for Spanish.  Out of a total of
75608 translations, our algorithm produces strings that match except for
whitespace differences 71223 times.  The 4548 mismatches are due to either
alternate ways of translating something or spelling mistakes in the translations.

Test Plan:
- npm test
- node index.js (runs regression_test.js)

Reviewers: jlfwong, james, chris

Subscribers: kitt

Differential Revision: https://phabricator.khanacademy.org/D21943