Update translate-assistant to support objects
Summary:
In Manticore we use translation items which are objects containing
the English string to be translated, the translated string, the translation
status, etc.  In order to work with these translation items this library
needs to be updated.  This diff adds the `groupObjects` function and
modifies the `suggest` method to take `items` instead of `englishStrs` along
with the `getEnglishStr` user supplied function which returns an English
string to be translated given an object (which in our case is a translation
item).

Test Plan: npm test

Reviewers: chris, james, jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D22560