Translate items that are only math, graphies, or widgets (take 2)
Summary:
We can provide suggestions for items that are only math, graphies, or widgets
by return the original English string as the translation because there is no
natural text in the string.  There are some edge cases that this doesn't cover,
e.g. natural language text that appears in math inside a `\text` command.  Also,
strings that contain multiple pieces of math will not be automatically translated.

Test Plan: npm tests

Reviewers: chris, james, jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D23062