Refactor translation-assistant so that suggest can handle English strings from multiple groups
Summary:
In order to simplify code in Manticore that's using translation-assistant
we'd like to make it more powerful so that it can handle cases where we'd
like to get suggestions for multiple strings in multiple different groups.
This case comes up when providing suggestions for previews.  This diff
introduces a Suggestor class which can be instantiated with a set of items
that are grouped along with any translation pairs if they exist.  The `suggest`
method can be passed a set of items that belong in one or more of these
groups.  The method will return the translation pairs `[item, translation]`.
If no translation pair existed for the group that `item` belongs to and the
`item` isn't strictly math, a graphie, or a widget, `translation` will be
null.

Test Plan: - npm test

Reviewers: chris, james, jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D23173