Differentiate **bold** text from non-bold text
Summary: The reason for this change is that we want to ensure that any English text that contain **bold** text will be translated to text that also contains **bold** text (same for non-bold text).  This is because the poentry linter checks for bolded text and we want to ensure that the suggestions from the translation assistant pass the linter.

Test Plan: npm test

Reviewers: jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D23602