Fix issues arising from trailing newline mismatches
Summary:
The translation assistant splits english and translated strings into
multiple lines using `\n\n` as a separate token.  In some cases this
token was appear at the end of the english string but not at the end
of the translated string.  This was causing issues in `populateTemplate`
(englishLines had more lines than template.lines) and in `stringToGroupKey`
(two groups were being formed when there should've been one).  In this
diff I trim whitespace from the end of the English strings to fix these issues.

Test Plan: npm test

Reviewers: jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D24637