Handle trailing newline in translated string to enable suggested smart translations
Summary:
In `populateTemplate` the English string is trimmed to remove trailing newlines,
while the translated string given to `createTemplate` (to create the template
used in suggesting a translation) was not trimmed.

This may result in mismatch between the number of text lines in the template and
the English string for which a smart translation is to be (and can be) suggested.

Also I fixed some tests that were not specifying an expected null (no translation
suggestion) value which resulted in test not asserting anything.

Before (gif):
{F982692}

After (gif):
{F982693}

https://app.asana.com/0/460596819119522/458361234110754

Test Plan:
 - In `webapp/third_party/javascript-khansrc/translation-assistant`
 - Run `npm run install && npm run build && npm run test`
 - In `webapp/`
 - Run `make serve`
 - Open `http://localhost:8080/translations/edit/tr/math/cc-eighth-grade-math/cc-8th-geometry/cc-8th-angles-between-lines/e/parallel_lines_2/by-pattern`
 - Verify `Add smart translations` button is enabled for a pattern that has some non-approved translations with trailing newline.

Reviewers: #contentplatform, kevinb, joshuan

Reviewed By: kevinb

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D40661