Do not offer translations when translated math notation does not fit.
e.g. when a translator forgets to properly translate the decimal point,
we will not offer STs. This might notify the translator of an actual
problem in the translations.

Otherwise, we might be hitting some special case, in which case it's
better to be conservative anyway and not offer STs.