FEATURE: Handle multiple math notations
JIRA: https://khanacademy.atlassian.net/browse/IC-662

Summary:
The current math translation system inside expects
that there is just one notation for a given operator for a given locale.
However, there might be different valid notations in some countries.
A typical example is \times and \cdot, the former is often used in lower
grades, but both are valid. We want to provide the translators the
flexibility to do both while still offering Smart Translations.

The proposed solution is a new function maybeTranslateMath(), which
decides which rules to apply by checking the template translations.
I went for the simplest heuristic I could think of. For example,
MAYBE_TIMES_AS_CDOT rule will be applied if:
1. Template translation contains \cdot at least once
anywhere in the translated text

AND

2. Template translation does not contain \times anywhere

It is unlikely that a single string would contain different notations.
If it does, we treat it as a translation mistake and ST will not be
offered.

Most importantly, this approach will allow us to handle cartesian
coordinates and intervals as well (will be done separately).
These are not possible to do atm because the US notation is the same for
both of these so we cannot determine how to translate them.

I was pleasantly surprised that the proposed change did not increase the
code complexity too much.

Obviously, this approach cannot work if the template is empty, which can
be the case for strings containing only math. In this case, ST suggestions
will keep the US notation and translators will need to translate
manually if needed. However, I did change the code so that the template
is applied even for the only-math strings if it is available. See the
tests for how that works.

Also, we suppose that within a group, the math strings contain the same
notation, although that might not always be the case. I can imagine an
exercise that includes some problems with division and some with
multiplication, while keeping the NL text the same. In such a case, the
ST suggestions would depend on which of the string was translated as a
template. Also, this will be the case for math-only strings, where it's
quite likely to happen.
If we find this behaviour problematic, we'd need to modify the grouping
algorithm to create different groups based on used math notation.
I'd rather avoid that since it would add a lot of complexity
that I tried to avoid here. In any case, this change is a clear
improvement and should benefit translators as is.

Test Plan:
Test various locales and multiplication notations in exercise:
https://www.khanacademy.org/devadmin/content/exercises/factoring_linear_binomials/1126131850/x85f73fd8684daede

Closes #24