Add .eslintignore and update README
If you happen to install khan-linter in this repo,
it would choke on the build files in `lib/` and block
the git push (via pre-push hook). Adding `.eslintignore`
fixes that.

I've also updated the deploy instructions in README
and explicitly mention the Khan Academy Dots plugin
that needs to be updated when `math-translator.js`
is updated.

Note that in webapp we no longer need the build files
in lib since TA is now imported via `webapp/package.json`
and part of the normal webpack build.
However, we still use `lib/math-translator.js` via git submodule
in the Khan Academy Dots repo.

Closes #16.

Issue: None

Test plan:
 - ka-lint should no longer blocks a push to Github.